requirements.txt
discord.py
yt-dlp
aiohttp
python-dotenv
PyNaCl